Solaris compat.

+++ hcore/compat.hxx
@@ -57,11 +57,11 @@ void* memrchr( void const*, int, int long );
 #endif /* not HAVE_DECL_SUN_LEN */
 
 #if ! defined( HAVE_GNU_GETHOSTBYNAME_R )
-int gethostbyname_r( char const*, struct hostent*, char*, size_t, struct hostent**, int* );
+int gethostbyname_r( char const*, struct hostent*, char*, long unsigned, struct hostent**, int* );
 #endif /* not HAVE_GNU_GETHOSTBYNAME_R */
 
 #if ! defined( HAVE_GNU_GETHOSTBYADDR_R )
-int gethostbyaddr_r( void const*, int, int, struct hostent*, char*, size_t, struct hostent**, int* );
+int gethostbyaddr_r( void const*, int, int, struct hostent*, char*, long unsigned, struct hostent**, int* );
 #endif /* not HAVE_GNU_GETHOSTBYADDR_R */
 
 #if ! defined( HAVE_POWL ) || ( HAVE_POWL == 0 )
